Update treatment-cost-malta-additional-2.html
Removal of the interim check your answers page when adding another treatment for the same patient with a different quota number. The user has the opportunity to check and change the details at the last step so it was just an additional step with no real value so can be removed.